212: Add reset functionality
