+++ lib/i18n/app_de.json
@@ -18,7 +18,8 @@
       "consent": "Verstanden",
       "settings": "Einstellungen",
       "openSettings": "Einstellungen öffnen",
-      "close": "Schließen"
+      "close": "Schließen",
+      "reset": "Filter zurücksetzen"
     }
   },
   "error": {
